+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export {default as useKakaoUser} from './useKakaoUser';
 export {default as useLocalUser} from './useLocalUser';
 export {default as useGetColumnGutter} from './useGetColumnGutter';
 export {default as useGetCurrentPosition} from './useGetCurrentPosition';
+export {default as useScrollBottomSheet} from './useScrollBottomSheet';
